+++ AAEmu.Game/Core/Packets/G2C/SCAuctionPostedPacket.cs
@@ -20,7 +20,7 @@ namespace AAEmu.Game.Core.Packets.G2C
             stream.Write(item.ItemID);
             stream.Write(item.ObjectID);
             stream.Write(item.Grade);
-            stream.Write(item.Flags);
+            stream.Write((byte)item.Flags);
             stream.Write(item.StackSize);
             stream.Write(item.DetailType);
             stream.Write(DateTime.Now);
